Share the surface/bbox tag attributes in wmtk::SurfaceTagAttributes
tetwild's and simwild's FaceAttributes and triwild's and simwild-2D's
EdgeAttributes were character-identical: two fields, m_is_surface_fs and
m_is_bbox_fs, plus reset() and merge(). Four copies become four aliases to one
struct in src/wmtk.

The 2D pair's unused `double tag;` goes with them. It was flagged
`// TODO: is this used?` in both copies; it is not -- nothing reads or writes it,
and the build confirms it.

107 lines removed for 35. No output change: ctest 107/107, all 53 registered
configs byte-identical apart from the two num_threads=10 tetwild configs and the
expected simwild-3D #t correction from the previous commit.
